requirements.txt
einops==0.8.0
natsort==8.4.0
numpy==1.24.3
opencv_python==4.9.0.80
Pillow==10.3.0
pytorch-cuda==11.6 
scikit_learn==1.3.2
SimpleITK==2.3.1
skimage==0.21.0
torch==1.13.0
torchvision==0.14.0
